Bugfix for kiwicoin.js
The main bug was separating the params with a semicolon instead of a comma inside the message var. This caused an "Unauthorized" error from Kiwi-Coin for any request with params - buy/sell/etc.

I've also changed the params from an object to an array, since Kiwi-Coin demands they be in a specific order, and you can't know for sure what order an object's keys will be in.

Lastly I removed the / 1000 from the nonce, since for me I can often be doing more than one per second - check a balance then immediately make a trade.